Implemented execution of Python apps. Closes #1
The python host provides the necessary libs to execute Python code.
These libs must be loaded by the Python app. To ensure that there is no
security issue the Python host must have the permission
"com.python.permission.PYTHONHOST".

It's a start, but I think that the starting process of the Python apps
is too complicated, not intuitive and takes too long. If anybody has
advice on that topic, please tell me at issue #9.