Remove apk bound libraries, use downloaded ones
Removed the old Python libraries that came with the apk.
This greatly reduces the apk size from over 30 Mb to 3.53 Mb!
Although the user needs to download the Python interpreter libraries and
modules, the final apk size with the same functionality as before is
11.41 Mb which is still less than half the size than before.
And if the user decides that he doesn't want the ssl, ctypes and bz2
module, only 9.89 Mb are used.

In addition to those changes, all other libraries who still are
installed within the apk (e.g the logging and the application library)
are now available for all CPU-ABIs.

Furthermore, a time estimation and a downloading animation will now be
displayed during download.
A preferences screen was also added, where the user (for now) only can
select the default Python version.

As a nice additional feature, typing "license()" in the interpreter now
shows the full license text.