Massively improve performance and memory usage of the Python terminal
To see the improvements to the Python terminal, try to run

for i in range(1000000):
i

on the Python terminal before this commit and afterwards. The memory
usage goes down to a Mb in comparison to the 4 to 6 Mb it took before.

A command history was also added, but is disabled for now until I come
up with a good way to access it.
Another major improvement:
The terminal input now reads all data from stdin that got enqueued while
it was inactive and takes it as input as soon as it becomes active
again. You can now type commands while you execute sleep().
Also the input properly works with soft input methods again.