Convert the terminal view to a Fragment
This is done in order to allow the Python App to use the fragment
provided by the Python Host. This will be implemented in the near
future.

The Python Host now has permission to access the external sdcard in
order to execute your scripts fom there and the command history can now
be accessed by horizontal swiping on the prompt line

Also fixed a crash that occurred when printing Utf-8 characters greater
than 127. This was caused by a bug in the jni NewStringUTF method which
incorrectly raises an exception.

Support for the PyOS_InputHook was added, which will be executed while
waiting for input.