Add the SDLWindow
Improved the Windowmanager, tabs can now be removed for real (the
default Android implementation doesn't really allow that for some
reason).

The current diretory of the python interpreter is now set to the
directory where the Python executable is stored.

Also, fixed a NPE when typing input before the interpreter was
connected.