Add PyFragment, cleanup the WindowManagerFragment
We had to get rid of the AppCombat Fragments, because they require the
activity to be a subclass of their special FragmentActivity and that did
not work with the classloading introduced by the Python App. My
implementation does not require the activity to be something special
(jet), but as a tradeoff, these Fragments do not receive onPause,
onRestore, etc...

The WindowManagerFragment (former TerminalWindowManagerFragment) doesn't
require a terminal tab to be at the first index anymore. It is now just
a tabbased manager for PyFragments.