Add AppInterpreter and reimplement the application lib
The AppInterpreter implements the new communication interface for the
Python apps. It allows the Python apps to reuse the Python hosts UI
elements.

The initial communication is still established via the
PythonAppCommunicationManager, but it mostly just gives the Python app
the path to the application lib, which the app has to load.
The application lib acts as a wrapper to the AppInterpreter and loads
the AppInterpreter into the process of the Python app.