Request showing the input focus on the output list, if the input EditText is not onscreen
This fixes #22. On newer Android versions, the software keyboard is not shown
if the view passed to `InputMethodManager#showSoftInput` is not on the screen.
Because the TerminalInput is only used when the readline module was installed,
the keyboard would never show up.

We now pass the output container as an argument instead, which forwards all input
to the TerminalInput anyways.